# storage/engine: Reduce one copy in FindSplitKey by reusing previous key buf