Currently, we copy the best split key inside pebbleIterator.FindSplitKey so
that it can be returned to the caller at the very end. At the same time, we
also track the previous key in a separate buffer to aid with size tracking.

This change optimizes findSplitKey in the most common case to just return the
prevKey buffer instead of making a copy. If we come across the harder case
of a good key candidate followed by a span of invalid split keys, it still
makes the double copy just in case.

Fixes #41602 .

Release note: None.