@itsbilal 2019-11-21
Not a very significant improvement, but it does show up:

```
v1040-wn-rt-c-20-71:cockroach bilal$ benchstat ./bench_findsplitkey_old ./bench_findsplitkey_new
name                                     old time/op    new time/op    delta
MVCCFindSplitKey_Pebble/valueSize=32-12    51.8ms ± 2%    49.3ms ± 4%  -4.84%  (p=0.000 n=14+15)

name                                     old speed      new speed      delta
MVCCFindSplitKey_Pebble/valueSize=32-12  1.30GB/s ± 2%  1.36GB/s ± 5%  +5.14%  (p=0.000 n=14+15)
```